Updates
Moved the creation of the Snake object to occur after the screen.tracer(0). This prevented the screen from flickering when it first opens. I changed the create_segment method so that a new segment is always created behind the last segment and it's heading is set in the same direction of the last segment. Because it no long needs coordinates to create a segment, the starting_position method was changed to call the create_segment three times without having to pass in coordinates. The START_POSITION constant was deleted.